+++ packages/app-store/ee/routing-forms/components/FormInputFields.tsx
@@ -1,8 +1,8 @@
 import { App_RoutingForms_Form } from "@prisma/client";
 import { Dispatch, SetStateAction } from "react";
 
+import { getQueryBuilderConfig } from "../lib/getQueryBuilderConfig";
 import isRouterLinkedField from "../lib/isRouterLinkedField";
-import { getQueryBuilderConfig } from "../pages/route-builder/[...appPages]";
 import { SerializableForm, Response } from "../types/types";
 
 type Props = {
