+++ packages/app-store/ee/routing-forms/lib/InitialConfig.tsx
@@ -1,0 +1,3 @@
+import QueryBuilderInitialConfig from "../components/react-awesome-query-builder/config/config";
+
+export const InitialConfig = QueryBuilderInitialConfig;
